Revert "SERVER-24363 cast mallinfo members to long long before adding"
This reverts commit 8900002b731358b0beedadb2ceb4e3156de402b6.

+++ src/mongo/util/processinfo_linux.cpp
@@ -451,8 +451,7 @@ void ProcessInfo::getExtraInfo(BSONObjBuilder& info) {
     struct mallinfo malloc_info =
         mallinfo();  // structure has same name as function that returns it. (see malloc.h)
     info.append("heap_usage_bytes",
-                static_cast<long long>(malloc_info.uordblks) /*main arena*/ +
-                    static_cast<long long>(malloc_info.hblkhd) /*mmap blocks*/);
+                malloc_info.uordblks /*main arena*/ + malloc_info.hblkhd /*mmap blocks*/);
     // docs claim hblkhd is included in uordblks but it isn't
 
     LinuxProc p(_pid);
